Refactoring by adding Timestampable trait
